fix: otimiza codigo
