+++ src/services/index.js
@@ -6,9 +6,11 @@ const app = express();
 
 app.listen(3000, () => console.log("Servidor iniciado na porta 3000"));
 
+const endpoints = require("./endpoints");
+
 module.exports = {
-  buzz: app.get("/Buzz", require("./buzz")),
-  fizz: app.get("/Fizz", require("./fizz")),
-  fizzBuzz: app.get("/FizzBuzz", require("./fizz-buzz")),
-  other: app.get("/Other", require("./other")),
+  buzz: app.get("/Buzz", endpoints.buzz),
+  fizz: app.get("/Fizz", endpoints.fizz),
+  fizzBuzz: app.get("/FizzBuzz", endpoints.fizzBuzz),
+  other: app.get("/Other", endpoints.other),
 };
